Added missing 16 bit cast for FFT size packet header field

+++ re/src/workers/fft_accel/data/fft_data.hpp
@@ -124,7 +124,7 @@ namespace sem::fft_accel::data {
         data_.at(2) = static_cast<std::byte>(native_packet.request_id());
         // 3rd byte currently unused, formerly used for CSR
         // TODO: Thoroughly investigate endianness of remaining fields to be serialized
-        uint16_t size_as_big_endian = boost::endian::native_to_big(native_packet.payload_data().size());
+        uint16_t size_as_big_endian = boost::endian::native_to_big(static_cast<uint16_t>(native_packet.payload_data().size()));
         reinterpret_cast<uint16_t&>(data_.at(4)) = size_as_big_endian;
 
         constexpr auto payload_byte_length = NumElements * sizeof(SampleType);
